requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
altair==5.0.1
ansi2html==1.8.0
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
async-timeout==4.0.2
attrs==23.1.0
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==6.0.0
blinker==1.6.2
Boruta==0.3
cachetools==5.3.1
catboost==1.2
category-encoders==2.6.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.6
cloudpickle==2.2.1
cmake==3.27.2
colorama==0.4.6
comm==0.1.3
contourpy==1.1.0
cycler==0.11.0
dash==2.11.1
dash-bootstrap-components==1.4.2
dash-core-components==2.0.0
dash-daq==0.5.0
dash-html-components==2.0.0
dash-table==5.0.0
datasets==2.12.0
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
detect-delimiter==0.1.1
dill==0.3.6
docopt==0.6.2
entrypoints==0.4
et-xmlfile==1.1.0
evaluate==0.4.0
#evdev==1.6.1
executing==1.2.0
fastjsonschema==2.17.1
filelock==3.12.0
Flask==2.2.5
fontawesome==5.10.1.post1
fonttools==4.41.0
fqdn==1.5.1
frozenlist==1.3.3
fsspec==2023.5.0
gitdb==4.0.10
GitPython==3.1.32
GoogleBard==1.1.1
graphviz==0.20.1
huggingface-hub==0.14.1
idna==3.4
imageio==2.31.1
imbalanced-learn==0.11.0
importlib-metadata==6.8.0
ipykernel==6.23.1
ipython==8.13.2
ipython-genutils==0.2.0
ipywidgets==7.7.2
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
joblib==1.3.1
Js2Py==0.74
jsonpointer==2.3
jsonschema==4.18.4
jsonschema-specifications==2023.7.1
jupyter-events==0.6.3
jupyter_client==8.2.0
jupyter_core==5.3.0
jupyter_server==2.5.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.4
kiwisolver==1.4.4
lazy_loader==0.3
lightgbm==4.0.0
lime==0.2.0.1
lit==16.0.6
llvmlite==0.40.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.2
matplotlib-inline==0.1.6
mdurl==0.1.2
missingpy==0.2.0
mistune==2.0.5
mpmath==1.3.0
mrmr-selection==0.2.8
multidict==6.0.4
multiprocess==0.70.14
nbclassic==1.0.0
nbclient==0.8.0
nbconvert==7.4.0
nbformat==5.8.0
nest-asyncio==1.5.6
networkx==3.1
nltk==3.8.1
notebook==6.5.4
notebook_shim==0.2.3
numba==0.57.1
numpy==1.24.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
#nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
#nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
openpyxl==3.1.2
packaging==23.1
pandas==2.0.3
pandocfilters==1.5.0
parso==0.8.3
patsy==0.5.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
pipreqs==0.4.13
pipwin==0.5.2
platformdirs==3.5.1
plotly==5.15.0
polars==0.16.9
prometheus-client==0.16.0
prompt-toolkit==3.0.38
protobuf==4.23.4
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==12.0.1
pycparser==2.21
pydeck==0.8.0
Pygments==2.15.1
pyjsparser==2.7.1
Pympler==1.0.1
pynput==1.7.6
pyparsing==3.0.9
PyPDF2==3.0.1
PyPrind==2.11.3
pyreadstat==1.2.2
pyrsistent==0.19.3
pySmartDL==1.3.4
python-dateutil==2.8.2
python-json-logger==2.0.7
python-xlib==0.33
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.4.1
PyYAML==6.0.1
pyzmq==25.0.2
referencing==0.30.0
regex==2023.5.5
requests==2.31.0
responses==0.18.0
retrying==1.3.4
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.4.2
rpds-py==0.9.2
scikit-image==0.21.0
scikit-learn==1.3.0
scipy==1.11.1
seaborn==0.12.2
Send2Trash==1.8.2
sentence-transformers==2.2.2
sentencepiece==0.1.99
setfit==0.7.0
shap==0.42.1
simplejson==3.19.1
six==1.16.0
sklearn-json==0.1.0
slicer==0.0.7
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.4.1
spss-converter==0.2.0
st-pages==0.4.3
stack-data==0.6.2
statsmodels==0.14.0
streamlit==1.25.0
sympy==1.12
tenacity==8.2.2
terminado==0.17.1
threadpoolctl==3.2.0
tifffile==2023.7.10
tinycss2==1.2.1
tk==0.1.0
tokenizers==0.13.3
toml==0.10.2
toolz==0.12.0
torch==2.0.1
torchvision==0.15.2
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
transformers==4.29.2
#triton==2.0.0
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==4.3.1
uri-template==1.2.0
urllib3==2.0.3
validator-collection==1.5.0
validators==0.20.0
watchdog==3.0.0
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.5.2
Werkzeug==2.2.3
widgetsnbextension==3.6.4
xgboost==1.7.6
xxhash==3.2.0
yarg==0.1.9
yarl==1.9.2
zipp==3.16.2
